refactor: добавит компонент main-content
